dont pull if archive exists

+++ scripts/get_tweets.R
@@ -21,20 +21,21 @@ read_latest <- function(){
 }
 
 # ---- get tweets ----
-jjstache_likes <- get_favorites("__jsta", n = 3000)
-jjstache_likes <- jjstache_likes[
-  order(jjstache_likes$created_at, decreasing = TRUE),]
-
-dt <- read_latest()
-dt2 <- jjstache_likes[1:
-                        which(jjstache_likes$status_id == dt[1, "status_id"]),]
-dt2 <- dplyr::select(dt2, -media_url, -mentions_screen_name,
-                     -mentions_user_id,
-                     -hashtags)
-
-res <- dplyr::bind_rows(dt2, dt)
-
-
 outfile <- file.path("data", paste0(Sys.Date(), "_jjstache_likes.rds"))
 print(outfile)
-saveRDS(res, outfile)
+if(!file.exists(outfile)){
+  jjstache_likes <- get_favorites("__jsta", n = 3000)
+  jjstache_likes <- jjstache_likes[
+    order(jjstache_likes$created_at, decreasing = TRUE),]
+
+  dt <- read_latest()
+  dt2 <- jjstache_likes[1:
+                          which(jjstache_likes$status_id == dt[1, "status_id"]),]
+  dt2 <- dplyr::select(dt2, -media_url, -mentions_screen_name,
+                       -mentions_user_id,
+                       -hashtags)
+
+  res <- dplyr::bind_rows(dt2, dt)
+
+  saveRDS(res, outfile)
+}
